google-bigquery/requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.2.3
-e repo-packages
-e workspace-packages
